represent code as postgresqlCode instead of expression/string
A postgresqlCode can contain plsqlStringVariables. In such cases
no subtree is added in IslandSqlDocument. Only if the
postgresqlCode consist of a single stringCodeElement subtrees
are produced. The reason is that plsqlStringVariable can
contain arbitrary code fragments which might be necessary to
successfully parse the code. A logic like in sqlName is not
applicable here.